chore(nestjs-gateway): move http interfaces to http boundary
