+++ packages/nestjs-gateway/src/mesh/http/express.gateway.ts
@@ -4,7 +4,7 @@ import type { RequestHandler }        from 'express'
 
 import type { GatewayModuleOptions }  from '../../module/interfaces.js'
 import type { GatewayGraphQLRuntime } from '../interfaces.js'
-import type { GatewayHttpBoundary }   from '../interfaces.js'
+import type { GatewayHttpBoundary }   from './interfaces.js'
 
 import { Injectable }                 from '@nestjs/common'
 import { HttpAdapterHost }            from '@nestjs/core'
