+++ packages/nestjs-gateway/src/mesh/index.ts
@@ -2,6 +2,7 @@ export *      from './errors/index.js'
 export *      from './graphql-mesh.handler.js'
 export *      from './graphql-mesh.config.js'
 export *      from './graphql.mesh.js'
+export type * from './http/interfaces.js'
 export *      from './http/express.gateway.js'
 export *      from './http/fastify.gateway.js'
 export type * from './interfaces.js'
