+++ packages/nestjs-gateway/src/mesh/interfaces.ts
@@ -1,10 +1,7 @@
-import type { ApolloServer }         from '@apollo/server'
-import type { MeshInstance }         from '@graphql-mesh/runtime'
-import type { GraphQLSchema }        from 'graphql'
-import type { IncomingMessage }      from 'node:http'
-import type { Socket }               from 'node:net'
-
-import type { GatewayModuleOptions } from '../module/interfaces.js'
+import type { ApolloServer }    from '@apollo/server'
+import type { MeshInstance }    from '@graphql-mesh/runtime'
+import type { GraphQLSchema }   from 'graphql'
+import type { IncomingMessage } from 'node:http'
 
 export type GatewayContext = Record<string, unknown>
 
@@ -22,15 +19,4 @@ export interface GatewayGraphQLRuntime {
   schema: GraphQLSchema
   contextBuilder: GatewayContextBuilder
   apolloServer: ApolloServer<GatewayContext>
-}
-
-export interface GatewayHttpServer {
-  on: (
-    event: 'upgrade',
-    handler: (req: IncomingMessage, socket: Socket, head: Buffer) => void
-  ) => void
-}
-
-export interface GatewayHttpBoundary {
-  register: (runtime: GatewayGraphQLRuntime, options: GatewayModuleOptions) => Promise<void>
 }
